+++ backend/app/models/item.py
@@ -6,32 +6,30 @@ from sqlalchemy import (
     func
 )
 from sqlalchemy.orm import Session, relationship
+from pgvector.sqlalchemy import Vector
 from .base import Base, LogicalDeleteMixin
 
 class Item(Base, LogicalDeleteMixin):
-    __tablename__ = "items"
+    __tablename__ = 'items'
 
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String, nullable=False)
     unit = Column(String, nullable=False, default='pcs')
     quantity = Column(Float, nullable=False, default=0.0)
     location = Column(String, nullable=True)
-    family_id = Column(Integer, ForeignKey("families.id"), nullable=False)
-    owner_id = Column(Integer, ForeignKey("users.id"), nullable=True)
+    family_id = Column(Integer, ForeignKey('families.id'), nullable=False)
+    owner_id = Column(Integer, ForeignKey('users.id'), nullable=True)
     check_interval_days = Column(Integer, nullable=True)
     last_checked_date = Column(DateTime, nullable=True)
     restock_threshold = Column(Float, nullable=True)
     notes = Column(String, nullable=True)
     raw_input = Column(String, nullable=True)
+    embedding = Column(Vector(1536), nullable=True)
 
-    # __table_args__ = (
-    #     UniqueConstraint("name", "unit", "location", "family_id", "owner_id", name="_item_uc"),
-    # )
-
-    family = relationship("Family", back_populates="items")
-    owner = relationship("User", back_populates="items", foreign_keys=[owner_id])
-    transactions = relationship("Transaction", back_populates="item")
-    tags = relationship("Tag", secondary="item_tags", back_populates="items")
+    family = relationship('Family', back_populates='items')
+    owner = relationship('User', back_populates='items', foreign_keys=[owner_id])
+    transactions = relationship('Transaction', back_populates='item')
+    tags = relationship('Tag', secondary='item_tags', back_populates='items')
 
     @classmethod
     def find_by_unique(cls, 
@@ -55,7 +53,20 @@ class Item(Base, LogicalDeleteMixin):
             family_id=family_id,
             owner_id=owner_id
         ).first()
-        
+    
+    def get_embedding_text(self) -> str:
+        '''Get a string representation of the item for embedding.'''
+        tag_names = ','.join([tag.name for tag in self.tags])
+        value_list = [
+            self.name,
+            self.unit,
+            self.location,
+            self.notes if self.notes else 'None',
+            'None' if tag_names == '' else tag_names
+        ]
+        text = '|||'.join(value_list)
+        return text.replace('\n', ' ').replace('\r', ' ')
+
 
 def get_unique_locations(db: Session, family_id: int):
     return (